models: update to reflect modified schema
This generalizes the schema to support multiple events, moves the team
size restriction into the event, and separates the sponsors and donors.
Furthermore, setup a mechanism to link the donor and sponsor to the
luminate identity via the `luminate_id`.  Normalise some of the names
additionally to make the schema seem more coherent and tie more
information back to the user.